Maintenance and warning cleanup

+++ Ganymede/src/Platform/Windows/WindowsInput.cpp
@@ -25,7 +25,7 @@ namespace Ganymede {
         auto window = static_cast<GLFWwindow*>( Application::Get().GetWindow().GetNativeWindow());
         double xPos, yPos;
         glfwGetCursorPos(window, &xPos, &yPos);
-        return {xPos, yPos};
+        return {(float)xPos, (float)yPos};
     }
 
     float WindowsInput::GetMouseXImpl() {
